pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "feedparser>=6.0.12",
    "flask>=3.1.2",
    "google-auth>=2.43.0",
    "gspread>=6.2.1",
    "lxml>=6.0.2",
    "openai>=2.7.2",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "schedule>=1.2.2",
]
